refactor(Whiteboard): move Whiteboard to its own component that only renders the drawing canvas and wrap in a parent Mime component that is in charge of rendering shapes from the store
