+++ browser/Root.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react'
 import Routes from './routes/Routes'
-import Whiteboard from './components/Whiteboard'
+import Mime from './components/Mime'
 
 const Root = () => {
 	return (
 		<section>
 			<Routes />
+			<Mime />
 		</section>
 	)
 }
